Fix customer operations
