+++ handler/add_order_item_to_cart_handler.go
@@ -2,7 +2,7 @@ package handler
 
 import (
 	"backend/database"
-	"backend/model"
+	. "backend/model"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
@@ -12,25 +12,30 @@ import (
 AddOrderItemToCartHandler is a function for gin to handle AddOrderItemToCart api
 */
 func AddOrderItemToCartHandler(c *gin.Context) {
-	addToCart := AddToCart
-	
+	var addToCart AddToCart
 
 	productID, err := strconv.Atoi(c.Query("p_id"))
 	if err != nil {
 		c.Status(400)
+	} else {
+		addToCart.ProductID = productID
 	}
 
 	cartID, err := strconv.Atoi(c.Query("cart_id"))
 	if err != nil {
 		c.Status(400)
+	} else {
+		addToCart.CartID = cartID
 	}
 
 	quantity, err := strconv.Atoi(c.Query("quantity"))
 	if err != nil {
 		c.Status(400)
+	} else {
+		addToCart.Quantity = quantity
 	}
 
-	code := database.AddOrderItemToCart(productID, cartID, quantity, database.SqlDb)
+	code := database.AddOrderItemToCart(&addToCart, database.SqlDb)
 
 	c.Status(code)
 }
