+++ handler/delete_order_item_in_cart_handler.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"backend/database"
+	. "backend/model"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
@@ -11,17 +12,23 @@ import (
 DeleteOrderItemToCartHandler is a function for gin to handle DeleteOrderItemToCart api
 */
 func DeleteOrderItemToCartHandler(c *gin.Context) {
+	var addToCart AddToCart
+
 	productID, err := strconv.Atoi(c.Query("p_id"))
 	if err != nil {
 		c.Status(400)
+	} else {
+		addToCart.ProductID = productID
 	}
 
 	cartID, err := strconv.Atoi(c.Query("cart_id"))
 	if err != nil {
 		c.Status(400)
+	} else {
+		addToCart.CartID = cartID
 	}
 
-	code := database.DeleteOrderItemInCart(productID, cartID, database.SqlDb)
+	code := database.DeleteOrderItemInCart(&addToCart, database.SqlDb)
 
 	c.Status(code)
 }
