+++ handler/get_order_items_in_cart_handler.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"backend/database"
+	. "backend/model"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
@@ -11,12 +12,16 @@ import (
 GetOrderItemsInCartHandler is a function for gin to handle GetOrderItemsInCart api
 */
 func GetOrderItemsInCartHandler(c *gin.Context) {
+	var addToCart AddToCart
+
 	cartID, err := strconv.Atoi(c.Query("cart_id"))
 	if err != nil {
 		c.Status(400)
+	} else {
+		addToCart.CartID = cartID
 	}
 
-	code, items := database.GetOrderItemsInCart(cartID, database.SqlDb)
+	code, items := database.GetOrderItemsInCart(&addToCart, database.SqlDb)
 
 	c.JSON(code, gin.H{
 		"items": items,
