+++ model/add_to_cart.go
@@ -5,18 +5,3 @@ type AddToCart struct {
 	CartID    int `json:"c_id"`
 	Quantity  int `json:"quantity"`
 }
-
-func (addToCart *AddToCart) SetProductID(productID int) error {
-	addToCart.ProductID = productID
-	return nil
-}
-
-func (addToCart *AddToCart) SetCartID(cartID int) error {
-	addToCart.CartID = cartID
-	return nil
-}
-
-func (addToCart *AddToCart) SetQuantity(quantity int) error {
-	addToCart.Quantity = quantity
-	return nil
-}
